Code organization + .Net 9 + New Work Item
